Correção no modelo cliente
